fix branch problem

+++ metrics/fid_batch/fid_evaluation.py
@@ -13,7 +13,7 @@ import lpips
         
 from .utils import create_batches
 
-def Fid_evatuation(dir_results, device, last_sample = False):
+def Fid_evatuation(dir_results, device, mmse_sample = True, last_sample = False, method = "ours"):
     """
     Calculate the Fréchet Inception Distance (FID) between the target and estimated images.
 
@@ -28,27 +28,31 @@ def Fid_evatuation(dir_results, device, last_sample = False):
     fid = pyiqa.create_metric("fid").to(device)
     
     dir_ref = os.path.join(dir_results, "ref")  
-    dir_mmse = os.path.join(dir_results, "mmse")
-    if os.path.exists(os.path.join(dir_results, "last")):
-        dir_last = os.path.join(dir_results, "last")
-        
-    # make patches
-    batch_dir_mmse = create_batches(dir_mmse, 4)
     batch_dir_ref = create_batches(dir_ref, 4)
-    batch_dir_last = create_batches(dir_last, 4) if last_sample else None
     
-    # computing fid
-    print("we want to compute the fid")
-    fid_mmse = fid(batch_dir_ref, batch_dir_mmse) 
-    fid_last = fid(batch_dir_ref, batch_dir_last) if last_sample else "Not computed" 
+    out = {}
+    if mmse_sample:
+        dir_mmse = os.path.join(dir_results, "mmse")
+        batch_dir_mmse = create_batches(dir_mmse, 4)
+        print("we want to compute the fid of the mmse estimate")
+        fid_mmse = fid(batch_dir_ref, batch_dir_mmse) 
+        out["fid_mmse"]= fid_mmse
+        shutil.rmtree(batch_dir_mmse)
+    if last_sample:
+        dir_last = os.path.join(dir_results, "last")
+        batch_dir_last = create_batches(dir_last, 4)
+        print("we want to compute the fid of the last estimate")
+        fid_last = fid(batch_dir_ref, batch_dir_last) 
+        out["fid_last"]= fid_last
+        shutil.rmtree(batch_dir_last)
+   
     print("fid computed")
     
-    # remove folder with patches
+
     shutil.rmtree(batch_dir_ref)
-    shutil.rmtree(batch_dir_mmse)
-    shutil.rmtree(batch_dir_last) if last_sample else None
+
     
-    return fid_mmse, fid_last
+    return out
